+++ lib/execute.ts
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-import { CommandInvocation } from "@atomist/automation-client/internal/invoker/Payload";
+import {
+    CommandInvocation,
+} from "@atomist/automation-client";
 import * as stringify from "json-stringify-safe";
 
 import { extractArgs } from "./command";